+++ src/convert.rs
@@ -17,3 +17,23 @@ impl AsFaceRef for &ttf_parser::Face<'_> {
         self
     }
 }
+
+/// Trait exposing mutable operations on a [`ttf_parser::Face`].
+pub trait FaceMut {
+    /// Sets a variation axis coordinate.
+    ///
+    /// See [`ttf_parser::Face::set_variation`].
+    fn set_variation(&mut self, axis: ttf_parser::Tag, value: f32) -> Option<()>;
+}
+impl FaceMut for ttf_parser::Face<'_> {
+    #[inline]
+    fn set_variation(&mut self, axis: ttf_parser::Tag, value: f32) -> Option<()> {
+        ttf_parser::Face::set_variation(self, axis, value)
+    }
+}
+impl FaceMut for &mut ttf_parser::Face<'_> {
+    #[inline]
+    fn set_variation(&mut self, axis: ttf_parser::Tag, value: f32) -> Option<()> {
+        ttf_parser::Face::set_variation(*self, axis, value)
+    }
+}