+++ src/owned.rs
@@ -87,6 +87,25 @@ impl crate::convert::AsFaceRef for &OwnedFace {
     }
 }
 
+impl crate::convert::FaceMut for OwnedFace {
+    fn set_variation(&mut self, axis: ttf_parser::Tag, value: f32) -> Option<()> {
+        unsafe {
+            let mut_ref = Pin::as_mut(&mut self.0);
+            let mut_inner = mut_ref.get_unchecked_mut();
+            match mut_inner.face.as_mut() {
+                Some(face) => face.set_variation(axis, value),
+                None => None,
+            }
+        }
+    }
+}
+impl crate::convert::FaceMut for &mut OwnedFace {
+    #[inline]
+    fn set_variation(&mut self, axis: ttf_parser::Tag, value: f32) -> Option<()> {
+        (*self).set_variation(axis, value)
+    }
+}
+
 // Face data in a `Vec` with a self-referencing `Face`.
 struct SelfRefVecFace {
     data: Vec<u8>,