+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased (0.15.2)
+* Add `FaceMut::set_variation` trait abstraction for calling mutable `Face::set_variation` via `OwnedFace`.
+* Use edition 2021.
+
 # 0.15.1
 * Add `OwnedFace::as_slice`, `OwnedFace::into_vec`.